evm: add interface to read and write EVM state (ENABLE/DISABLE/FIX).
The state of EVM can be change by writing to the <securityfs>/evm_state,
and can be check by reading from the interface.
To change the state of EVM write:
0 - for disabling
1 - for enabling
2 - for FIX mode

Notice that evm have to be initialized before you use the interface.

To enable this option you have to set a proper option in Kconfig.

Change-Id: Id68546fcae3c32087b85314da863a0c5bd9e1a94

Signed-off-by: Janusz Kozerski <j.kozerski@samsung.com>