packaging: better conditional build of trace/event
There was when building linux v4.0-rc3 before that patch.
The libraries /lib/traceevent/plugins/*.so were not found.

This is fixed now.

When tuning the build flag 'trace_supported' it also appeared
that this flag is broken and produce a packaging error if set to 0.
The solution used here after some experiment is to remove the
files built during installation. That isn't the best but it does the job.

Change-Id: Ib18f60da890c946c852b5eb4947d64362868de85
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>